Add install step to fix compat between kerl & recent ubuntu versions
Issue fixed in [`kerl` PR #521](https://github.com/kerl/kerl/pull/521) and its related PRs.

Alternatively, the default kerl version could be updated from 4.0.0 to 4.1.1 which would largely remove the need for this change. This would leave the compatibility issue documented only in GH issues as it is currently, with new installs avoiding the issue simply by following the install instructions and using the defaults.